requirements.txt
customtkinter==5.0.3
Pillow==9.3.0
pygame==2.1.2
